shows 3 search results + show more button #374
