+++ web/router.ex
@@ -31,7 +31,7 @@ defmodule App.Router do
     post "/feedback", FeedbackController, :post
     get "/crisis", CrisisController, :index
     post "/search", HomepageController, :search
-    get "/:page", InfoController, :index
+    get "/:page", StaticController, :index
     # Default route - will match any page - must stay at bottom
   end
