+++ web/static/js/show_more.js
@@ -1,18 +1,17 @@
-selectAll('#results > div > div').forEach(function (el, i) {
-  if(el.id.substring(0, 9) === 'resource_' && i > 3) {
+selectAll('.resource').forEach(function (el, i) {
+  if(el.id.substring(0, 9) === 'resource_' && i > 2) {
     toggleClasses(el, ['dn']);
   }
 });
 function handle_see_more(e) {
-  selectAll('#results > div > div').forEach(function (el, i) {
+  selectAll('.resource').forEach(function (el, i) {
     if(el.id.substring(0, 9) === 'resource_' && el.className.indexOf('dn') > -1) {
       toggleClasses(el, ['dn']);
     }
   });
   toggleClasses(select('#see_more'), ['dn-important']);
 }
-if (selectAll('#results > div > div').length > 3) {
+if (selectAll('.resource').length > 3) {
   toggleClasses(select('#see_more'), ['dn-important']);
   select('#see_more').addEventListener('click', handle_see_more);
 }
-
